# PR: Unscoped enums access for PyQt6 and other missing PyQt6 compatibility changes