--- a/qtpy/QtCore.py
+++ b/qtpy/QtCore.py
@@ -8,18 +8,19 @@
 """
 Provides QtCore classes and functions.
 """
-
 from . import PYQT6, PYQT5, PYSIDE2, PYSIDE6, PythonQtError
 
+
 if PYQT6:
+    from PyQt6 import QtCore
     from PyQt6.QtCore import *
     from PyQt6.QtCore import pyqtSignal as Signal
     from PyQt6.QtCore import QT_VERSION_STR as __version__
-
     QCoreApplication.exec_ = QCoreApplication.exec
     QEventLoop.exec_ = QEventLoop.exec
     QThread.exec_ = QThread.exec
-
+    from .enums_compat import promote_enums
+    promote_enums(QtCore)
 elif PYQT5:
     from PyQt5.QtCore import *
     from PyQt5.QtCore import pyqtSignal as Signal