--- a/qtpy/QtGui.py
+++ b/qtpy/QtGui.py
@@ -8,16 +8,17 @@
 """
 Provides QtGui classes and functions.
 """
-import warnings
-
 from . import PYQT6, PYQT5, PYSIDE2, PYSIDE6, PythonQtError
 
 
 if PYQT6:
+    from PyQt6 import QtGui 
     from PyQt6.QtGui import *
     QDrag.exec_ = QDrag.exec
     QGuiApplication.exec_ = QGuiApplication.exec
     QTextDocument.print_ = QTextDocument.print
+    from .enums_compat import promote_enums
+    promote_enums(QtGui)
 elif PYQT5:
     from PyQt5.QtGui import *
 elif PYSIDE2: