--- a/qtpy/QtGui.py
+++ b/qtpy/QtGui.py
@@ -14,9 +14,13 @@ from . import PYQT6, PYQT5, PYSIDE2, PYSIDE6, PythonQtError
 if PYQT6:
     from PyQt6 import QtGui 
     from PyQt6.QtGui import *
+
+    # Map missing/renamed methods
     QDrag.exec_ = QDrag.exec
     QGuiApplication.exec_ = QGuiApplication.exec
     QTextDocument.print_ = QTextDocument.print
+
+    # Allow unscoped access for enums inside the QtGui module
     from .enums_compat import promote_enums
     promote_enums(QtGui)
 elif PYQT5: