--- a/qtpy/QtWidgets.py
+++ b/qtpy/QtWidgets.py
@@ -8,23 +8,25 @@
 """
 Provides widget classes and functions.
 """
-
 from . import PYQT5, PYQT6, PYSIDE2, PYSIDE6, PythonQtError
-from ._patch.qheaderview import introduce_renamed_methods_qheaderview
+
 
 if PYQT6:
+    from PyQt6 import QtWidgets
     from PyQt6.QtWidgets import *
     from PyQt6.QtGui import QAction, QActionGroup, QShortcut
     from PyQt6.QtOpenGLWidgets import QOpenGLWidget
     QTextEdit.setTabStopWidth = QTextEdit.setTabStopDistance
     QTextEdit.tabStopWidth = QTextEdit.tabStopDistance
+    QTextEdit.print_ = QTextEdit.print
     QPlainTextEdit.setTabStopWidth = QPlainTextEdit.setTabStopDistance
     QPlainTextEdit.tabStopWidth = QPlainTextEdit.tabStopDistance
+    QPlainTextEdit.print_ = QPlainTextEdit.print
     QApplication.exec_ = QApplication.exec
     QDialog.exec_ = QDialog.exec
     QMenu.exec_ = QMenu.exec
-    QTextEdit.print_ = QTextEdit.print
-    QPlainTextEdit.print_ = QPlainTextEdit.print
+    from .enums_compat import promote_enums
+    promote_enums(QtWidgets)
 elif PYQT5:
     from PyQt5.QtWidgets import *
 elif PYSIDE6: