--- a/qtpy/QtWidgets.py
+++ b/qtpy/QtWidgets.py
@@ -16,6 +16,8 @@ if PYQT6:
     from PyQt6.QtWidgets import *
     from PyQt6.QtGui import QAction, QActionGroup, QShortcut
     from PyQt6.QtOpenGLWidgets import QOpenGLWidget
+
+    # Map missing/renamed methods
     QTextEdit.setTabStopWidth = QTextEdit.setTabStopDistance
     QTextEdit.tabStopWidth = QTextEdit.tabStopDistance
     QTextEdit.print_ = QTextEdit.print
@@ -25,6 +27,8 @@ if PYQT6:
     QApplication.exec_ = QApplication.exec
     QDialog.exec_ = QDialog.exec
     QMenu.exec_ = QMenu.exec
+
+    # Allow unscoped access for enums inside the QtWidgets module
     from .enums_compat import promote_enums
     promote_enums(QtWidgets)
 elif PYQT5:
@@ -33,6 +37,8 @@ elif PYSIDE6:
     from PySide6.QtWidgets import *
     from PySide6.QtGui import QAction, QActionGroup, QShortcut
     from PySide6.QtOpenGLWidgets import QOpenGLWidget
+
+    # Map missing/renamed methods
     QTextEdit.setTabStopWidth = QTextEdit.setTabStopDistance
     QTextEdit.tabStopWidth = QTextEdit.tabStopDistance
     QPlainTextEdit.setTabStopWidth = QPlainTextEdit.setTabStopDistance