--- a/qtpy/QtWidgets.py
+++ b/qtpy/QtWidgets.py
@@ -31,6 +31,7 @@ if PYQT6:
     # Allow unscoped access for enums inside the QtWidgets module
     from .enums_compat import promote_enums
     promote_enums(QtWidgets)
+    del QtWidgets
 elif PYQT5:
     from PyQt5.QtWidgets import *
 elif PYSIDE6: