--- a/qtpy/enums_compat.py
+++ b/qtpy/enums_compat.py
@@ -2,7 +2,7 @@
 # Copyright © 2009- The Spyder Development Team
 # Licensed under the terms of the MIT License
 """
-Compatibility functions for scoped and unscoped enums access
+Compatibility functions for scoped and unscoped enum access.
 """
 from . import PYQT6
 
@@ -36,15 +36,15 @@ if PYQT6:
         Taken from:
         https://github.com/pyqtgraph/pyqtgraph/blob/pyqtgraph-0.12.1/pyqtgraph/Qt.py#L331-L377 
         """
-        class_names = [x for x in dir(module) if x.startswith('Q')]
+        class_names = [name for name in dir(module) if name.startswith('Q')]
         for class_name in class_names:
             klass = getattr(module, class_name)
             if not isinstance(klass, sip.wrappertype):
                 continue
-            attrib_names = [x for x in dir(klass) if x[0].isupper()]
+            attrib_names = [name for name in dir(klass) if name[0].isupper()]
             for attrib_name in attrib_names:
                 attrib = getattr(klass, attrib_name)
                 if not isinstance(attrib, enum.EnumMeta):
                     continue
-                for e in attrib:
-                    setattr(klass, e.name, e)
+                for enum_obj in attrib:
+                    setattr(klass, enum_obj.name, enum_obj)