--- a/qtpy/sip.py
+++ b/qtpy/sip.py
@@ -1,0 +1,15 @@
+
+#
+# Copyright © 2009- The Spyder Development Team
+#
+# Licensed under the terms of the MIT License
+# (see LICENSE.txt for details)
+
+from . import PYQT6, PYQT5,PythonQtError
+
+if PYQT6:
+    from PyQt6.sip import *
+elif PYQT5:
+    from PyQt5.sip import *
+else:
+    raise PythonQtError('No Qt bindings could be found')