--- a/qtpy/sip.py
+++ b/qtpy/sip.py
@@ -1,4 +1,3 @@
-
 #
 # Copyright © 2009- The Spyder Development Team
 #
@@ -12,4 +11,5 @@ if PYQT6:
 elif PYQT5:
     from PyQt5.sip import *
 else:
-    raise PythonQtError('No Qt bindings could be found')
+    raise PythonQtError(
+        'Currently selected Qt binding does not support this module')