--- a/qtpy/tests/test_qtcore.py
+++ b/qtpy/tests/test_qtcore.py
@@ -1,5 +1,5 @@
 import pytest
-from qtpy import PYQT5, PYQT6, PYSIDE2, QtCore
+from qtpy import PYQT5, PYQT6, PYSIDE2, PYQT_VERSION, QtCore
 
 """Test QtCore."""
 
@@ -25,3 +25,18 @@ def test_QtCore_SignalInstance():
     instance = ClassWithSignal()
 
     assert isinstance(instance.signal, QtCore.SignalInstance)
+
+
+@pytest.mark.skipif(PYQT5 and PYQT_VERSION.startswith('5.9'),
+                    reason="A specific setup with at least sip 4.9.9 is needed for PyQt5 5.9.*"
+                           "to work with scoped enum access")
+def test_enum_access():
+    """
+    Test scoped and unscoped enum access for qtpy.QtCore.*.
+    """
+    assert QtCore.QAbstractAnimation.Stopped == QtCore.QAbstractAnimation.State.Stopped
+    assert QtCore.QEvent.ActionAdded == QtCore.QEvent.Type.ActionAdded
+    assert QtCore.Qt.AlignLeft == QtCore.Qt.AlignmentFlag.AlignLeft
+    assert QtCore.Qt.Key_Return == QtCore.Qt.Key.Key_Return
+    assert QtCore.Qt.transparent == QtCore.Qt.GlobalColor.transparent
+    assert QtCore.Qt.Widget == QtCore.Qt.WindowType.Widget