--- a/qtpy/tests/test_qtcore.py
+++ b/qtpy/tests/test_qtcore.py
@@ -10,8 +10,6 @@ def test_qtmsghandler():
     assert QtCore.qInstallMessageHandler is not None
 
 
-@pytest.mark.skipif(not (PYQT5 or PYSIDE2),
-                    reason="Targeted to PyQt5 or PySide2")
 def test_DateTime_toPython():
     """Test QDateTime.toPython"""
     assert QtCore.QDateTime.toPython is not None