Part of #233