@ccordoba12 2021-11-01
@dalthviz, I was thinking that in order to avoid the penalty hit involved in loading all those symbols at startup, we could use lazy loading instead. In other words, if the symbol is requested then we load it automatically. If not, it won't be loaded at all.

These are some resources that could help you with that:

* https://levelup.gitconnected.com/python-trick-lazy-module-loading-df9b9dc111af
* https://anvil.works/blog/lazy-modules
* https://snarky.ca/lazy-importing-in-python-3-7/
* https://github.com/mnmelo/lazy_import

I think the third one is the most promising. Unfortunately, that requires Python 3.7, but we could make Qtpy depend on that version given that 3.6 is about to reach end of life.