@CAM-Gerlach 2021-11-02
Given we've supported 3.6 up to this point with the QtPy 2.0 release, 3.6 still sees some significant use (e.g. its still the default Python on many CIs, distros, etc), and the bindings we support all still support it (particularly the Qt6 ones, which would be impossible to use with any version of QtPy on Python 3.6 otherwise, and in fact Qt6 + Py36 would actually see QtPy downgraded to a mutually incompatible version), if we did go ahead with this approach, it would make sense to guard the lazy imports behind a version check, or try-except, IMO.