@dalthviz 2021-11-02
> I was thinking that in order to avoid the penalty hit involved in loading all those symbols at startup, we could use lazy loading instead.

Then we will need to precalculate a map of possible import calls without the use of enumeration classes? I'm not totally sure to understand here the use of lazy loading.

The penalty involved with what is here is more related to the automatic detection of the enums and setting the values at module level than the import it self I think (that's why I was experimenting with `QtCore.Qt` enums by having before hand the enum name classes).