@CAM-Gerlach 2021-11-02
@dalthviz Just in case you aren't already aware and it would be helpful, `python -X importtime -c "import qtpy"` (or more specifically, `-c "import qtpy.modulename"`) in Python 3.7+ profiles the time taken to import qtpy and each dep, arranged as a reverse tree.