@dalthviz 2021-11-03
Just in case doing some benchmarking tests using the `timeit` module and @CAM-Gerlach suggestion:

## `timeit` testing

### `from qtpy import QtCore, QtWidgets, QtGui`

* With `promote_enums` being call:
```
(pyqt6) C:\Users\dalth>python -m timeit "from qtpy import QtCore, QtWidgets, QtGui"
1 loop, best of 5: 1.2 usec per loop
:0: UserWarning: The test results are likely unreliable. The worst time (7.1 usec) was more than four times slower than the best time (1.2 usec).
```
* Without `promote_enums` being call:

```
(pyqt6) C:\Users\dalth\Desktop>python -m timeit "from qtpy import QtCore, QtWidgets, QtGui"
200000 loops, best of 5: 949 nsec per loop
```


### `import qtpy.QtCore`

* With `promote_enums` being call:
```
python -m timeit "import qtpy.QtCore"
1000000 loops, best of 5: 297 nsec per loop
```

* Without `promote_enums` being call:

```
(pyqt6) C:\Users\dalth\Desktop>python -m timeit "import qtpy.QtCore"
1000000 loops, best of 5: 292 nsec per loop
```

### `import qtpy.QtWidgets`
* With `promote_enums` being call:
```
(pyqt6) C:\Users\dalth\Desktop>python -m timeit "import qtpy.QtWidgets"
1000000 loops, best of 5: 295 nsec per loop

```

* Without `promote_enums` being call:

```
(pyqt6) C:\Users\dalth\Desktop>python -m timeit "import qtpy.QtWidgets"
1000000 loops, best of 5: 290 nsec per loop
```

### `import qtpy.QtGui`
* With `promote_enums` being call:
```
(pyqt6) C:\Users\dalth\Desktop>python -m timeit "import qtpy.QtGui"
1000000 loops, best of 5: 295 nsec per loop
```

* Without `promote_enums` being call:

```
(pyqt6) C:\Users\dalth\Desktop>python -m timeit "import qtpy.QtGui"
1000000 loops, best of 5: 292 nsec per loop
```


## `importtime` testing

### `from qtpy import QtCore, QtWidgets, QtGui`

* With `promote_enums` being call:

<details>

```
import time: self [us] | cumulative | imported package
import time:       321 |        321 |   _io
import time:       143 |        143 |   marshal
import time:       259 |        259 |   nt
import time:       126 |        126 |   winreg
import time:      2399 |       3247 | _frozen_importlib_external
import time:       108 |        108 |   time
import time:       396 |        503 | zipimport
import time:       185 |        185 |     _codecs
import time:      1228 |       1413 |   codecs
import time:      1204 |       1204 |   encodings.aliases
import time:      2846 |       5462 | encodings
import time:       780 |        780 | encodings.utf_8
import time:       757 |        757 | encodings.cp1252
import time:        70 |         70 | _signal
import time:        66 |         66 |     _abc
import time:       885 |        951 |   abc
import time:      1204 |       2154 | io
import time:       102 |        102 |       _stat
import time:       907 |       1008 |     stat
import time:      1771 |       1771 |     _collections_abc
import time:       903 |        903 |       genericpath
import time:      1384 |       2287 |     ntpath
import time:      1810 |       6874 |   os
import time:       832 |        832 |   _sitebuiltins
import time:      1186 |       1186 |     pywin32_system32
import time:      2693 |       3878 |   pywin32_bootstrap
import time:       606 |        606 |   sitecustomize
import time:       643 |        643 |   usercustomize
import time:      5051 |      17881 | site
import time:       927 |        927 |       packaging.__about__
import time:      1275 |       2202 |     packaging
import time:       216 |        216 |       itertools
import time:       808 |        808 |       keyword
import time:       108 |        108 |         _operator
import time:      1052 |       1159 |       operator
import time:       899 |        899 |       reprlib
import time:       112 |        112 |       _collections
import time:      3178 |       6369 |     collections
import time:       862 |        862 |         types
import time:      1473 |       2335 |       enum
import time:       107 |        107 |         _sre
import time:      1041 |       1041 |           sre_constants
import time:      1448 |       2488 |         sre_parse
import time:      1893 |       4487 |       sre_compile
import time:       109 |        109 |         _functools
import time:      1725 |       1833 |       functools
import time:        96 |         96 |       _locale
import time:       904 |        904 |       copyreg
import time:      2448 |      12101 |     re
import time:       999 |        999 |     warnings
import time:       816 |        816 |       collections.abc
import time:      1234 |       1234 |       contextlib
import time:      3025 |       5074 |     typing
import time:       938 |        938 |     packaging._structures
import time:      7548 |      35229 |   packaging.version
import time:       103 |        103 |       errno
import time:      1206 |       1206 |       signal
import time:       851 |        851 |         _weakrefset
import time:      1567 |       2417 |       threading
import time:       733 |        733 |       fcntl
import time:        76 |         76 |       msvcrt
import time:       124 |        124 |       _winapi
import time:      2357 |       7014 |     subprocess
import time:      4001 |      11015 |   platform
import time:      1088 |       1088 |   qtpy._version
import time:       681 |        681 |     PyQt5
import time:       152 |        832 |   PyQt5.QtCore
import time:      1122 |       1122 |         importlib
import time:       743 |        743 |           importlib._abc
import time:      1224 |       1967 |         importlib.util
import time:       690 |        690 |         importlib.machinery
import time:      1083 |       1083 |         weakref
import time:      1799 |       6660 |       pkgutil
import time:      3046 |       9705 |     PyQt6
import time:        81 |         81 |       atexit
import time:      1477 |       1558 |     PyQt6.sip
import time:     15103 |      26365 |   PyQt6.QtCore
import time:       139 |        139 |     PyQt6.QtDataVisualization
import time:       906 |       1045 |   qtpy.QtDataVisualization
import time:      3472 |      79043 | qtpy
import time:       572 |        572 |     qtpy.sip
import time:       886 |       1458 |   qtpy.enums_compat
import time:     78344 |      79801 | qtpy.QtCore
import time:      4753 |       4753 |     PyQt6.QtGui
import time:      8579 |      13331 |   PyQt6.QtWidgets
import time:      2436 |       2436 |     PyQt6.QtOpenGL
import time:      4527 |       6963 |   PyQt6.QtOpenGLWidgets
import time:     57861 |      78154 | qtpy.QtWidgets
import time:     38523 |      38523 | qtpy.QtGui

```

</details>

* Without `promote_enums` being call:

<details>

```
import time: self [us] | cumulative | imported package
import time:       207 |        207 |   _io
import time:        64 |         64 |   marshal
import time:       216 |        216 |   nt
import time:       115 |        115 |   winreg
import time:      1124 |       1724 | _frozen_importlib_external
import time:       111 |        111 |   time
import time:      2148 |       2258 | zipimport
import time:       187 |        187 |     _codecs
import time:      1169 |       1356 |   codecs
import time:      1175 |       1175 |   encodings.aliases
import time:      2388 |       4918 | encodings
import time:       889 |        889 | encodings.utf_8
import time:       797 |        797 | encodings.cp1252
import time:        73 |         73 | _signal
import time:        88 |         88 |     _abc
import time:      1497 |       1585 |   abc
import time:      1288 |       2872 | io
import time:        95 |         95 |       _stat
import time:       805 |        899 |     stat
import time:      1789 |       1789 |     _collections_abc
import time:       784 |        784 |       genericpath
import time:      1544 |       2328 |     ntpath
import time:      1850 |       6864 |   os
import time:       844 |        844 |   _sitebuiltins
import time:      1211 |       1211 |     pywin32_system32
import time:      2933 |       4143 |   pywin32_bootstrap
import time:       577 |        577 |   sitecustomize
import time:       620 |        620 |   usercustomize
import time:      5054 |      18100 | site
import time:       941 |        941 |       packaging.__about__
import time:      1315 |       2256 |     packaging
import time:       220 |        220 |       itertools
import time:       797 |        797 |       keyword
import time:       106 |        106 |         _operator
import time:      1110 |       1216 |       operator
import time:       881 |        881 |       reprlib
import time:       109 |        109 |       _collections
import time:      3261 |       6482 |     collections
import time:       892 |        892 |         types
import time:      1556 |       2448 |       enum
import time:       118 |        118 |         _sre
import time:      1020 |       1020 |           sre_constants
import time:      1681 |       2701 |         sre_parse
import time:      2915 |       5733 |       sre_compile
import time:       103 |        103 |         _functools
import time:      1681 |       1784 |       functools
import time:       100 |        100 |       _locale
import time:       879 |        879 |       copyreg
import time:      2603 |      13545 |     re
import time:       984 |        984 |     warnings
import time:      1125 |       1125 |       collections.abc
import time:      1344 |       1344 |       contextlib
import time:      3231 |       5699 |     typing
import time:       909 |        909 |     packaging._structures
import time:      7377 |      37249 |   packaging.version
import time:       111 |        111 |       errno
import time:      1212 |       1212 |       signal
import time:       855 |        855 |         _weakrefset
import time:      1656 |       2511 |       threading
import time:       762 |        762 |       fcntl
import time:        81 |         81 |       msvcrt
import time:       131 |        131 |       _winapi
import time:      2511 |       7317 |     subprocess
import time:      3841 |      11157 |   platform
import time:      1020 |       1020 |   qtpy._version
import time:       649 |        649 |     PyQt5
import time:       119 |        767 |   PyQt5.QtCore
import time:       981 |        981 |         importlib
import time:       763 |        763 |           importlib._abc
import time:      1235 |       1998 |         importlib.util
import time:       699 |        699 |         importlib.machinery
import time:      1060 |       1060 |         weakref
import time:      1924 |       6660 |       pkgutil
import time:      3203 |       9863 |     PyQt6
import time:        79 |         79 |       atexit
import time:      1756 |       1834 |     PyQt6.sip
import time:     15151 |      26847 |   PyQt6.QtCore
import time:       137 |        137 |     PyQt6.QtDataVisualization
import time:       972 |       1109 |   qtpy.QtDataVisualization
import time:      2751 |      80897 | qtpy
import time:       615 |        615 |     qtpy.sip
import time:       901 |       1515 |   qtpy.enums_compat
import time:      6006 |       7521 | qtpy.QtCore
import time:      4728 |       4728 |     PyQt6.QtGui
import time:      9097 |      13824 |   PyQt6.QtWidgets
import time:      2488 |       2488 |     PyQt6.QtOpenGL
import time:      4104 |       6591 |   PyQt6.QtOpenGLWidgets
import time:    382741 |     403156 | qtpy.QtWidgets
import time:      9334 |       9334 | qtpy.QtGui
```

</details>